+++ src/main/java/com/upem/fr/ressource/CalendrierRessource.java
@@ -44,21 +44,29 @@ public class CalendrierRessource {
 
     @GetMapping("calendrier/{id}")
     public Optional<Calendrier> getOne(@PathVariable Long id) {
-        try {
-            return calendrierService.getOne(id);
-        }catch (NotFoundException e){
+        if(!calendrierService.getOne(id).isPresent())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  id  +  " inconnu");
-        }
+
+        return calendrierService.getOne(id);
+
     }
 
     @DeleteMapping("calendrier/{id}")
     public void delete(@PathVariable Long id) {
+        if(!calendrierService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  id  +  " inconnu");
+
         calendrierService.delete(id);
+
     }
 
     @PostMapping("calendrier/{id}")
     public Calendrier update(@PathVariable Long id, @Valid @RequestBody Calendrier calendrier) {
+        if(!calendrierService.getOne(id).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  id  +  " inconnu");
+
         return calendrierService.update(id, calendrier);
+
     }
